Use Tooltip as generic component
